requirements.txt
azure-mgmt-compute==19.0.0
azure-mgmt-resource==1.16.0
azure-identity==1.5.0
azure-storage-blob==12.8.0
